separated button data from slider button data file to its own file and imported it
