+++ libs/models/schema/ui/buttons/src/lib/button.interface.ts
@@ -1,0 +1,10 @@
+export interface ButtonData 
+{
+  text: string
+  
+  bgColor?: string; // Background colour
+  color?: string; // text color
+  
+  hoverBgColor?: string;
+  hoverColor?: string;
+}
