+++ libs/models/schema/ui/buttons/src/lib/slider-button-data.interface.ts
@@ -1,3 +1,5 @@
+import { ButtonData } from "./button.interface";
+
 export interface SliderButtonData extends ButtonData
 {
   text: string;
@@ -13,15 +15,4 @@ export interface SliderButtonData extends ButtonData
   hoverBorderColor?: string;
   hoverIconColor?: string;
   hoverIconBackgroundColor?: string;
-}
-
-export interface ButtonData 
-{
-  text: string
-  
-  bgColor?: string; // Background colour
-  color?: string; // text color
-  
-  hoverBgColor?: string;
-  hoverColor?: string;
 }
